Introduced two new classes: LocalPod and Sandbox.
The Sandbox represents the entire contents of the POD_ROOT (normally SOURCE_ROOT/Pods). A LocalPod represents a pod that has been installed within the Sandbox.

These two classes can be used as better homes for various pieces of logic currently spread throughout the installation process and provide a better API for working with the contents of this directory.

So far I've just refactored the downloaders to use them.

+++ lib/cocoapods/downloader.rb
@@ -6,29 +6,31 @@ module Pod
 
     extend Executable
 
-    def self.for_source(pod_root, source)
-      options = source.dup
+    def self.for_pod(pod)
+      options = pod.specification.source.dup
       if url = options.delete(:git)
-        Git.new(pod_root, url, options)
+        Git.new(pod, url, options)
       elsif url = options.delete(:hg)
-        Mercurial.new(pod_root, url, options)
+        Mercurial.new(pod, url, options)
       elsif url = options.delete(:svn)
-        Subversion.new(pod_root, url, options)
+        Subversion.new(pod, url, options)
       else
-        raise "Unsupported download strategy `#{source.inspect}'."
+        raise "Unsupported download strategy `#{options.inspect}'."
       end
     end
 
-    attr_reader :pod_root, :url, :options
+    attr_reader :pod, :url, :options
 
-    def initialize(pod_root, url, options)
-      @pod_root, @url, @options = pod_root, url, options
+    def initialize(pod, url, options)
+      @pod, @url, @options = pod, url, options
     end
 
     def clean(clean_paths = [])
+      return unless clean_paths
+      
       clean_paths.each do |path|
         path.rmtree
-      end if clean_paths
+      end
     end
   end
 end
